requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.1.0
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chex==0.1.86
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
etils==1.12.1
filelock==3.17.0
flax==0.8.4
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
grpcio==1.70.0
humanize==4.12.1
hydra-core==1.3.2
idna==3.10
importlib-resources==6.5.2
jax==0.4.26
jaxlib==0.4.26
jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lazy-loader==0.4
librosa==0.10.2
lightning==2.2.4
lightning-utilities==0.14.0
llvmlite==0.44.0
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opt-einsum==3.4.0
optax==0.2.2
orbax-checkpoint==0.6.4
packaging==24.2
pandas==2.2.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
propcache==0.3.0
protobuf==6.30.0
pycparser==2.22
pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2025.1
pyyaml==6.0.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==76.0.0
six==1.17.0
soundfile==0.12.1
soxr==0.5.0.post1
sympy==1.13.3
synthax==0.2.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorstore==0.1.72
threadpoolctl==3.5.0
toolz==1.0.0
torch==2.3.0
torchaudio==2.3.0
torchmetrics==1.6.2
torchvision==0.18.0
tqdm==4.66.4
triton==2.3.0
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
